# contracts/openapi.v0.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Svac.PublicApi | v0",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/auth/email-code": {
      "post": {
        "tags": [
          "AuthEndpoints"
        ],
        "operationId": "PostAuthEmailCode",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthEmailCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
    "/v1/auth/session": {
      "post": {
        "tags": [
          "AuthEndpoints"
        ],
        "operationId": "PostAuthSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCreated"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "tags": [
          "AuthEndpoints"
        ],
        "operationId": "PostAuthRefresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCreated"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthEmailCodeRequest": {
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "AuthRefreshRequest": {
        "required": [
          "refreshToken"
        ],
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "AuthSessionRequest": {
        "required": [
          "email",
          "code"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "code": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "OpaqueId": {
        "type": "string",
        "description": "A prefixed ULID (e.g. \"usr_01H...\"). No raw Guid/uuid ever crosses the API (SLICE_S1_CONTRACT.md §1b).",
        "format": "opaque-id"
      },
      "Problem": {
        "required": [
          "type",
          "title",
          "status",
          "messageKey",
          "correlationId"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "messageKey": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SessionCreated": {
        "required": [
          "accessToken",
          "accessExpiresAt",
          "refreshToken",
          "accountId"
        ],
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "accessExpiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "refreshToken": {
            "type": "string"
          },
          "accountId": {
            "$ref": "#/components/schemas/OpaqueId"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "description": "Opaque, server-side, revocable session access token (SLICE_S3_CONTRACT.md §1b). Prefix \"sst_\", minted by /v1/signup/complete, /v1/auth/session, and /v1/auth/refresh. Header only: `Authorization: Bearer sst_...` — never a cookie or query parameter at S3. Carries zero decodable claims (never a JWT); revoked/expired/unknown tokens resolve to the anonymous actor server-side, never a distinct 401 oracle.",
        "scheme": "bearer",
        "bearerFormat": "sst_<256-bit-random>"
      }
    }
  },
...
}